adapt quests to new design system
