+++ packages/keychain/src/components/quests/types.ts
@@ -1,0 +1,5 @@
+export enum QuestState {
+  Incomplete,
+  Claimable,
+  Complete,
+}
